header method changes

+++ api/controllers/ChatController.php
@@ -96,12 +96,12 @@ class ChatController
         if ($res) {
           $this->getHeaders();
           echo json_encode(['message' => 'Message sent successfully', 'status' => 200, 'ok' => 1]);
-        }else{
+        } else {
           throw new Exception('Error sending message');
         }
       } catch (Exception $e) {
         $this->getHeaders();
-        echo json_encode(['message' => $e->getMessage(), 'status' => 500, 'ok' => 0, 'error'=>$this->db->error]);
+        echo json_encode(['message' => $e->getMessage(), 'status' => 500, 'ok' => 0, 'error' => $this->db->error]);
       }
     } else {
       $this->getHeaders();
@@ -161,19 +161,29 @@ class ChatController
         $this->getHeaders();
         echo json_encode(['message' => $e->getMessage(), 'status' => 500, 'ok' => 0]);
       }
-    } elseif ($method == 'DELETE') {
+    } elseif ($method == 'POST') {
       /* delete single message */
-      try {
-        $message_id = $_GET['message_id'];
-        $sql = 'DELETE FROM messages WHERE id = ?';
-        $stmt = $this->db->prepare($sql);
-        $stmt->bind_param('s', $message_id);
-        $stmt->execute();
+      if (!isset($_POST['message_id'])) {
+        throw new Exception('Message id not set');
+      }
+      $method = $_POST['method'];
+      if ($method == 'DELETE') {
+        try {
+          $message_id = $_POST['message_id'];
+          $sql = 'DELETE FROM messages WHERE id = ?';
+          $stmt = $this->db->prepare($sql);
+          $stmt->bind_param('s', $message_id);
+          $stmt->execute();
+          $this->getHeaders();
+          echo json_encode(['message' => 'Message deleted successfully', 'status' => 200, 'ok' => 1]);
+        } catch (Exception $e) {
+          $this->getHeaders();
+          echo json_encode(['message' => $e->getMessage(), 'status' => 500, 'ok' => 0]);
+        }
+      } else {
         $this->getHeaders();
-        echo json_encode(['message' => 'Message deleted successfully', 'status' => 200, 'ok' => 1]);
-      } catch (Exception $e) {
-        $this->getHeaders();
-        echo json_encode(['message' => $e->getMessage(), 'status' => 500, 'ok' => 0]);
+        echo json_encode(['message' => 'Method not allowed', 'status' => 405, 'ok' => 0]);
+        exit;
       }
     } else {
       $this->getHeaders();
